+++ src/assets/static/js/wptelegram--p2tg-classic.00dd4837.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('jquery'), 'version' => '153fd144d20471db430ed4e54ff19489');