+++ src/includes/Utils.php
@@ -95,7 +95,7 @@ class Utils {
 	 */
 	public static function fitler_rest_errors( $response, $handler, $request ) {
 
-		$matches_route    = 0 === strpos( ltrim( $request->get_route(), '/' ), RESTController::NAMESPACE );
+		$matches_route    = 0 === strpos( ltrim( $request->get_route(), '/' ), RESTController::REST_NAMESPACE );
 		$is_invalid_param = is_wp_error( $response ) && 'rest_invalid_param' === $response->get_error_code();
 
 		if ( ! $is_invalid_param || ! $matches_route ) {