+++ src/modules/proxy/Main.php
@@ -2,8 +2,6 @@
 /**
  * The file that defines the module
  *
- * A class definition that includes attributes and functions used across the module
- *
  * @link       https://manzoorwani.dev
  * @since      1.0.0
  *
@@ -26,14 +24,6 @@ use WPTelegram\Core\modules\BaseModule;
 class Main extends BaseModule {
 
 	/**
-	 * The single instance of the class.
-	 *
-	 * @since 3.0.0
-	 * @var   Main $instance The instance.
-	 */
-	protected static $instance = null;
-
-	/**
 	 * Register all of the hooks.
 	 *
 	 * @since    1.0.0
@@ -41,7 +31,7 @@ class Main extends BaseModule {
 	 */
 	protected function define_on_active_hooks() {
 
-		$handler = new ProxyHandler( $this );
+		$handler = ProxyHandler::instance();
 
 		add_action( 'wptelegram_bot_api_remote_request_init', [ $handler, 'configure_proxy' ] );