+++ src/modules/proxy/ProxyHandler.php
@@ -61,24 +61,29 @@ class ProxyHandler extends BaseClass {
 	 * @since  2.0.8
 	 */
 	public function setup_proxy() {
-
-		$proxy_method = $this->module->options()->get( 'proxy_method' );
+		$proxy_method = $this->module()->options()->get( 'proxy_method' );
 
 		if ( ! $proxy_method ) {
 			return;
 		}
+		$cf_worker_url     = $this->module()->options()->get( 'cf_worker_url' );
+		$google_script_url = $this->module()->options()->get( 'google_script_url' );
 
-		$script_url = $this->module->options()->get( 'google_script_url' );
+		if ( 'cf_worker' === $proxy_method && ! empty( $cf_worker_url ) ) {
 
-		if ( 'php_proxy' === $proxy_method ) {
+			// set URL.
+			add_filter( 'wptelegram_bot_api_base_url', [ __CLASS__, 'set_cf_worker_url' ], 10, 1 );
+
+		} elseif ( 'google_script' === $proxy_method && ! empty( $google_script_url ) ) {
+
+			// setup Google Script args.
+			add_filter( 'wptelegram_bot_api_remote_post_args', [ __CLASS__, 'google_script_request_args' ], 30, 2 );
+			// set URL.
+			add_filter( 'wptelegram_bot_api_request_url', [ __CLASS__, 'google_script_request_url' ], 30, 1 );
+
+		} elseif ( 'php_proxy' === $proxy_method ) {
 
 			$this->setup_php_proxy();
-
-		} elseif ( 'google_script' === $proxy_method && $script_url ) {
-			// setup Google Script args.
-			add_filter( 'wptelegram_bot_api_remote_post_args', [ __CLASS__, 'google_script_request_args' ], 20, 2 );
-			// set URL.
-			add_filter( 'wptelegram_bot_api_request_url', [ __CLASS__, 'google_script_request_url' ], 20, 1 );
 		}
 	}
 
@@ -96,13 +101,18 @@ class ProxyHandler extends BaseClass {
 			'password' => '',
 		];
 
-		// Met the values from settings/defaults.
+		// Get the values from settings/defaults.
 		foreach ( $defaults as $key => $value ) {
-			self::$proxy[ $key ] = $this->module->options()->get( 'proxy_' . $key, $value );
+			self::$proxy[ $key ] = $this->module()->options()->get( 'proxy_' . $key, $value );
 		}
 
-		// Modify curl.
-		add_action( 'http_api_curl', [ __CLASS__, 'modify_http_api_curl' ], 10, 3 );
+		// host is required.
+		if ( empty( self::$proxy['host'] ) ) {
+			return;
+		}
+
+		// modify curl.
+		add_action( 'http_api_curl', [ __CLASS__, 'modify_http_api_curl' ], 11, 3 );
 	}
 
 	/**
@@ -110,8 +120,8 @@ class ProxyHandler extends BaseClass {
 	 *
 	 * @return array
 	 */
-	private static function get_proxy() {
-		return (array) apply_filters( 'wptelegram_bot_api_curl_proxy', self::$proxy );
+	private static function get_php_proxy() {
+		return (array) apply_filters( 'wptelegram_php_proxy_options', self::$proxy );
 	}
 
 	/**
@@ -122,12 +132,15 @@ class ProxyHandler extends BaseClass {
 	 *
 	 * @since 1.0.0
 	 *
-	 * @param resource $handle  The cURL handle (passed by reference).
-	 * @param array    $r       The HTTP request arguments.
-	 * @param string   $url     The request URL.
+	 * @param \CurlHandle $handle  The cURL handle (passed by reference).
+	 * @param array       $r       The HTTP request arguments.
+	 * @param string      $url     The request URL.
+	 *
+	 * @return void
 	 */
 	public static function modify_http_api_curl( &$handle, $r, $url ) {
 
+		// set proxy only when using Telegram API directly without CF worker.
 		$bot_api_url = Client::BASE_URL;
 		$user_link   = 'https://t.me/';
 
@@ -135,17 +148,18 @@ class ProxyHandler extends BaseClass {
 
 		$to_telegram = preg_match( $pattern, $url );
 
-		$by_wptelegram = ( isset( $r['headers']['wptelegram_bot'] ) && $r['headers']['wptelegram_bot'] );
+		$by_wptelegram = ! empty( $r['headers']['wptelegram_bot'] );
 
 		// if the request is sent to Telegram by WP Telegram.
 		if ( $to_telegram && $by_wptelegram ) {
 
-			foreach ( self::get_proxy() as $option => $value ) {
-				${'proxy_' . $option} = apply_filters( "wptelegram_bot_api_curl_proxy_{$option}", $value );
+			foreach ( self::get_php_proxy() as $option => $value ) {
+				${'proxy_' . $option} = apply_filters( "wptelegram_php_proxy_option_{$option}", $value );
 			}
 
 			if ( ! empty( $proxy_host ) && ! empty( $proxy_port ) ) {
-				// phpcs:disable
+
+				// phpcs:disable -- Using cURL functions is highly discouraged.
 				curl_setopt( $handle, CURLOPT_PROXYTYPE, constant( $proxy_type ) );
 				curl_setopt( $handle, CURLOPT_PROXY, $proxy_host );
 				curl_setopt( $handle, CURLOPT_PROXYPORT, $proxy_port );
@@ -153,10 +167,9 @@ class ProxyHandler extends BaseClass {
 				if ( ! empty( $proxy_username ) && ! empty( $proxy_password ) ) {
 					$authentication = $proxy_username . ':' . $proxy_password;
 					curl_setopt( $handle, CURLOPT_PROXYAUTH, CURLAUTH_ANY );
-
 					curl_setopt( $handle, CURLOPT_PROXYUSERPWD, $authentication );
 				}
-				// phpcs:enable
+				// phpcs:enable -- Using cURL functions is highly discouraged.
 			}
 		}
 	}
@@ -164,10 +177,12 @@ class ProxyHandler extends BaseClass {
 	/**
 	 * Set Google Script args.
 	 *
+	 * @since  1.7.8
+	 *
 	 * @param array   $args    The request args.
 	 * @param Request $request The request object.
 	 *
-	 * @since  1.7.8
+	 * @return array
 	 */
 	public static function google_script_request_args( $args, $request ) {
 
@@ -182,12 +197,26 @@ class ProxyHandler extends BaseClass {
 	}
 
 	/**
-	 * Set Google Script URL
+	 * Set Google Script URL.
 	 *
 	 * @since  1.7.8
+	 *
+	 * @return string
 	 */
 	public static function google_script_request_url() {
 
 		return WPTG()->options()->get_path( 'proxy.google_script_url' );
+	}
+
+	/**
+	 * Set Cloudflare worker URL.
+	 *
+	 * @since  3.1.0
+	 *
+	 * @return string
+	 */
+	public static function set_cf_worker_url() {
+
+		return untrailingslashit( WPTG()->options()->get_path( 'proxy.cf_worker_url' ) ) . '/bot';
 	}
 }