+++ src/changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project are documented in this file.
 
 ## Unreleased
 
+## [3.1.0 - 2021-05-30](https://github.com/manzoorwanijk/wptelegram/releases/tag/v3.1.0)
+
+### Enhancements
+
+- Added CloudFlare Proxy support 🚀
+- Added `{post_type}` and `{post_type_label}` macros
+
 ## [3.0.11 - 2021-05-7](https://github.com/manzoorwanijk/wptelegram/releases/tag/v3.0.11)
 
 ### Bug fixes